avoid compiler warning
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/multicastBLD/current@1234 ececa60f-9f63-0410-97c7-ea05888eb260

+++ multicastTestApp/src/multicastTestSub.cpp
@@ -217,7 +217,7 @@ int MulticastTestSub::readPV(const char *sVariableName, int iBufferSize, void* p
     	return 1;
     }
 	
-	if ( (unsigned int) pBuffer & 0x3 != 0  )
+	if ( (unsigned long) pBuffer & 0x3 != 0  )
 	{
 		printf( "readPV(): Buffer should be aligned in 32 bits boundaries\n" );
     	return 2;
